+++ GGJ20/Assets/Scripts/Work/TaskManagers/HeatingTaskManager.cs
@@ -38,7 +38,7 @@ public class HeatingTaskManager : TaskManagerBase
         }
     }
 
-    public override int GetOffsetFromTarget()
+    public override float GetOffsetFromTarget()
     {
         return Mathf.Abs(targetHeat - currentHeat);
     }
